Refactoring for DB widgets
